requirements.txt
accelerate==0.25.0
asttokens==2.4.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
diffusers==0.11.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.9.0
fonttools==4.46.0
fsspec==2023.12.1
ftfy==6.1.3
huggingface-hub==0.19.4
idna==3.4
importlib-metadata==7.0.0
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.0
numpy==1.24.1
opencv-python==4.8.1.78
packaging==23.2
parso==0.8.3
Pillow==9.3.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.10.3
requests==2.28.1
safetensors==0.4.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tokenizers==0.15.0
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.35.2
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.12
zipp==3.17.0
